+++ lnd/lnrpc/rest-annotations.yaml
@@ -181,6 +181,13 @@ http:
       post: "/pkt/v1/sendfrom"
       body: "*"   
 
+    # metaservice.proto
+    - selector: lnrpc.MetaService.ChangePassword
+      post: "/pkt/v1/changepassword"
+      body: "*"
+    - selector: lnrpc.MetaService.GetInfo2
+      get: "/pkt/v1/getinfo2"
+
     # walletunlocker.proto
     - selector: lnrpc.WalletUnlocker.GenSeed
       get: "/v1/genseed"
@@ -189,9 +196,6 @@ http:
       body: "*"
     - selector: lnrpc.WalletUnlocker.UnlockWallet
       post: "/v1/unlockwallet"
-      body: "*"
-    - selector: lnrpc.WalletUnlocker.ChangePassword
-      post: "/v1/changepassword"
       body: "*"
     - selector: lnrpc.WalletUnlocker.CreateWallet
       post: "/pkt/v1/createwallet"
