+++ lnd/lnrpc/walletunlocker.proto
@@ -61,12 +61,6 @@ service WalletUnlocker {
     */
     rpc UnlockWallet (UnlockWalletRequest) returns (UnlockWalletResponse);
 
-    /* lncli: `changepassword`
-    ChangePassword changes the password of the encrypted wallet. This will
-    automatically unlock the wallet database if successful.
-    */
-    rpc ChangePassword (ChangePasswordRequest) returns (ChangePasswordResponse);
-
     /* The create command is used to initialize an lnd wallet from scratch for
 	the very first time. */
     rpc CreateWallet (CreateWalletRequest) returns (CreateWalletResponse);
@@ -202,51 +196,6 @@ message UnlockWalletRequest {
     bool stateless_init = 5;
 }
 message UnlockWalletResponse {
-}
-
-message ChangePasswordRequest {
-    /*
-    current_password should be the current valid passphrase used to unlock the
-    daemon. When using REST, this field must be encoded as base64.
-    */
-    bytes current_password = 1;
-
-    /* to support old pld created wallet we need to pass the pub password as well */
-    bytes current_pub_password = 2;
-
-    /*
-    new_password should be the new passphrase that will be needed to unlock the
-    daemon. When using REST, this field must be encoded as base64.
-    */
-    bytes new_password = 3;
-
-    bytes new_pub_password = 4;
-
-    /*
-    stateless_init is an optional argument instructing the daemon NOT to create
-    any *.macaroon files in its filesystem. If this parameter is set, then the
-    admin macaroon returned in the response MUST be stored by the caller of the
-    RPC as otherwise all access to the daemon will be lost!
-    */
-    bool stateless_init = 5;
-
-    /*
-    new_macaroon_root_key is an optional argument instructing the daemon to
-    rotate the macaroon root key when set to true. This will invalidate all
-    previously generated macaroons.
-    */
-    bool new_macaroon_root_key = 6;
-}
-message ChangePasswordResponse {
-    /*
-    The binary serialized admin macaroon that can be used to access the daemon
-    after rotating the macaroon root key. If both the stateless_init and
-    new_macaroon_root_key parameter were set to true, this is the ONLY copy of
-    the macaroon that was created from the new root key and MUST be stored
-    safely by the caller. Otherwise a copy of this macaroon is also persisted on
-    disk by the daemon, together with other macaroon files.
-    */
-    bytes admin_macaroon = 1;
 }
 
 message CreateWalletRequest {
