+++ lnd/lnrpc/walletunlocker.swagger.json
@@ -42,33 +42,6 @@
         ]
       }
     },
-    "/v1/changepassword": {
-      "post": {
-        "summary": "lncli: `changepassword`\nChangePassword changes the password of the encrypted wallet. This will\nautomatically unlock the wallet database if successful.",
-        "operationId": "ChangePassword",
-        "responses": {
-          "200": {
-            "description": "A successful response.",
-            "schema": {
-              "$ref": "#/definitions/lnrpcChangePasswordResponse"
-            }
-          }
-        },
-        "parameters": [
-          {
-            "name": "body",
-            "in": "body",
-            "required": true,
-            "schema": {
-              "$ref": "#/definitions/lnrpcChangePasswordRequest"
-            }
-          }
-        ],
-        "tags": [
-          "WalletUnlocker"
-        ]
-      }
-    },
     "/v1/genseed": {
       "get": {
         "summary": "GenSeed is the first method that should be used to instantiate a new lnd\ninstance. This method allows a caller to generate a new aezeed cipher seed\ngiven an optional passphrase. If provided, the passphrase will be necessary\nto decrypt the cipherseed to expose the internal wallet seed.",
@@ -172,50 +145,6 @@
         "multi_chan_backup": {
           "$ref": "#/definitions/lnrpcMultiChanBackup",
           "description": "A multi-channel backup that covers all open channels currently known to\nlnd."
-        }
-      }
-    },
-    "lnrpcChangePasswordRequest": {
-      "type": "object",
-      "properties": {
-        "current_password": {
-          "type": "string",
-          "format": "byte",
-          "description": "current_password should be the current valid passphrase used to unlock the\ndaemon. When using REST, this field must be encoded as base64."
-        },
-        "current_pub_password": {
-          "type": "string",
-          "format": "byte",
-          "title": "to support old pld created wallet we need to pass the pub password as well"
-        },
-        "new_password": {
-          "type": "string",
-          "format": "byte",
-          "description": "new_password should be the new passphrase that will be needed to unlock the\ndaemon. When using REST, this field must be encoded as base64."
-        },
-        "new_pub_password": {
-          "type": "string",
-          "format": "byte"
-        },
-        "stateless_init": {
-          "type": "boolean",
-          "format": "boolean",
-          "title": "stateless_init is an optional argument instructing the daemon NOT to create\nany *.macaroon files in its filesystem. If this parameter is set, then the\nadmin macaroon returned in the response MUST be stored by the caller of the\nRPC as otherwise all access to the daemon will be lost!"
-        },
-        "new_macaroon_root_key": {
-          "type": "boolean",
-          "format": "boolean",
-          "description": "new_macaroon_root_key is an optional argument instructing the daemon to\nrotate the macaroon root key when set to true. This will invalidate all\npreviously generated macaroons."
-        }
-      }
-    },
-    "lnrpcChangePasswordResponse": {
-      "type": "object",
-      "properties": {
-        "admin_macaroon": {
-          "type": "string",
-          "format": "byte",
-          "description": "The binary serialized admin macaroon that can be used to access the daemon\nafter rotating the macaroon root key. If both the stateless_init and\nnew_macaroon_root_key parameter were set to true, this is the ONLY copy of\nthe macaroon that was created from the new root key and MUST be stored\nsafely by the caller. Otherwise a copy of this macaroon is also persisted on\ndisk by the daemon, together with other macaroon files."
         }
       }
     },
