+++ lnd/walletunlocker/service.go
@@ -12,9 +12,7 @@ import (
 	"github.com/pkt-cash/pktd/lnd/chanbackup"
 	"github.com/pkt-cash/pktd/lnd/lncfg"
 	"github.com/pkt-cash/pktd/lnd/lnrpc"
-	"github.com/pkt-cash/pktd/lnd/lnwallet"
 	"github.com/pkt-cash/pktd/lnd/lnwallet/btcwallet"
-	"github.com/pkt-cash/pktd/lnd/macaroons"
 	"github.com/pkt-cash/pktd/pktlog/log"
 	"github.com/pkt-cash/pktd/pktwallet/wallet"
 	"github.com/pkt-cash/pktd/pktwallet/wallet/seedwords"
@@ -435,191 +433,6 @@ func (u *UnlockerService) UnlockWallet0(ctx context.Context,
 		select {
 		case <-u.MacResponseChan:
 			return &lnrpc.UnlockWalletResponse{}, nil
-
-		case <-ctx.Done():
-			return nil, ErrUnlockTimeout.Default()
-		}
-
-	case <-ctx.Done():
-		return nil, ErrUnlockTimeout.Default()
-	}
-}
-
-func (u *UnlockerService) ChangePassword(ctx context.Context,
-	in *lnrpc.ChangePasswordRequest) (*lnrpc.ChangePasswordResponse, error) {
-	res, err := u.ChangePassword0(ctx, in)
-	return res, er.Native(err)
-}
-
-// ChangePassword changes the password of the wallet and sends the new password
-// across the UnlockPasswords channel to automatically unlock the wallet if
-// successful.
-func (u *UnlockerService) ChangePassword0(ctx context.Context,
-	in *lnrpc.ChangePasswordRequest) (*lnrpc.ChangePasswordResponse, er.R) {
-
-	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
-	if u.walletPath != "" {
-		netDir = u.walletPath
-	}
-	loader := wallet.NewLoader(u.netParams, netDir, u.walletFile, u.noFreelistSync, 0)
-
-	// First, we'll make sure the wallet exists for the specific chain and
-	// network.
-	walletExists, err := loader.WalletExists()
-	if err != nil {
-		return nil, err
-	}
-
-	if !walletExists {
-		return nil, er.New("wallet not found")
-	}
-
-	publicPw := []byte(wallet.InsecurePubPassphrase)
-	if in.CurrentPubPassword != nil {
-		publicPw = in.CurrentPubPassword
-	}
-	privatePw := in.CurrentPassword
-	newPubPw := []byte(wallet.InsecurePubPassphrase)
-	if in.NewPubPassword != nil {
-		newPubPw = in.NewPubPassword
-	}
-	// If the current password is blank, we'll assume the user is coming
-	// from a --noseedbackup state, so we'll use the default passwords.
-	if len(in.CurrentPassword) == 0 {
-		publicPw = lnwallet.DefaultPublicPassphrase
-		privatePw = lnwallet.DefaultPrivatePassphrase
-	}
-
-	// Make sure the new password meets our constraints.
-	if err := ValidatePassword(in.NewPassword); err != nil {
-		return nil, err
-	}
-	if in.NewPubPassword != nil {
-		if err := ValidatePassword(in.NewPassword); err != nil {
-			return nil, err
-		}
-	}
-
-	// Load the existing wallet in order to proceed with the password change.
-	w, err := loader.OpenExistingWallet(publicPw, false)
-	if err != nil {
-		return nil, err
-	}
-
-	// Now that we've opened the wallet, we need to close it in case of an
-	// error. But not if we succeed, then the caller must close it.
-	orderlyReturn := false
-	defer func() {
-		if !orderlyReturn {
-			_ = loader.UnloadWallet()
-		}
-	}()
-
-	// Before we actually change the password, we need to check if all flags
-	// were set correctly. The content of the previously generated macaroon
-	// files will become invalid after we generate a new root key. So we try
-	// to delete them here and they will be recreated during normal startup
-	// later. If they are missing, this is only an error if the
-	// stateless_init flag was not set.
-	if in.NewMacaroonRootKey || in.StatelessInit {
-		for _, file := range u.macaroonFiles {
-			err := os.Remove(file)
-			if err != nil && !in.StatelessInit {
-				return nil, er.Errorf("could not remove "+
-					"macaroon file: %v. if the wallet "+
-					"was initialized stateless please "+
-					"add the --stateless_init "+
-					"flag", err)
-			}
-		}
-	}
-
-	// Attempt to change both the public and private passphrases for the
-	// wallet. This will be done atomically in order to prevent one
-	// passphrase change from being successful and not the other.
-	err = w.ChangePassphrases(
-		publicPw, newPubPw, privatePw, in.NewPassword,
-	)
-	if err != nil {
-		return nil, er.Errorf("unable to change wallet passphrase: "+
-			"%v", err)
-	}
-
-	// Check if macaroonFiles is populated, if not it's due to noMacaroon flag is set
-	//so we do not need the service
-	if len(u.macaroonFiles) > 0 {
-		// The next step is to load the macaroon database, change the password
-		// then close it again.
-		// Attempt to open the macaroon DB, unlock it and then change
-		// the passphrase.
-		macaroonService, err := macaroons.NewService(
-			netDir, "lnd", in.StatelessInit,
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		err = macaroonService.CreateUnlock(&privatePw)
-		if err != nil {
-			closeErr := macaroonService.Close()
-			if closeErr != nil {
-				return nil, er.Errorf("could not create unlock: %v "+
-					"--> follow-up error when closing: %v", err,
-					closeErr)
-			}
-			return nil, err
-		}
-		err = macaroonService.ChangePassword(privatePw, in.NewPassword)
-		if err != nil {
-			closeErr := macaroonService.Close()
-			if closeErr != nil {
-				return nil, er.Errorf("could not change password: %v "+
-					"--> follow-up error when closing: %v", err,
-					closeErr)
-			}
-			return nil, err
-		}
-
-		// If requested by the user, attempt to replace the existing
-		// macaroon root key with a new one.
-		if in.NewMacaroonRootKey {
-			err = macaroonService.GenerateNewRootKey()
-			if err != nil {
-				closeErr := macaroonService.Close()
-				if closeErr != nil {
-					return nil, er.Errorf("could not generate "+
-						"new root key: %v --> follow-up error "+
-						"when closing: %v", err, closeErr)
-				}
-				return nil, err
-			}
-		}
-
-		err = macaroonService.Close()
-		if err != nil {
-			return nil, er.Errorf("could not close macaroon service: %v",
-				err)
-		}
-	}
-	// Finally, send the new password across the UnlockPasswords channel to
-	// automatically unlock the wallet.
-	walletUnlockMsg := &WalletUnlockMsg{
-		Passphrase:    in.NewPassword,
-		Wallet:        w,
-		StatelessInit: in.StatelessInit,
-		UnloadWallet:  loader.UnloadWallet,
-	}
-	select {
-	case u.UnlockMsgs <- walletUnlockMsg:
-		// We need to read from the channel to let the daemon continue
-		// its work and to get the admin macaroon. Once the response
-		// arrives, we directly forward it to the client.
-		orderlyReturn = true
-		select {
-		case adminMac := <-u.MacResponseChan:
-			return &lnrpc.ChangePasswordResponse{
-				AdminMacaroon: adminMac,
-			}, nil
 
 		case <-ctx.Done():
 			return nil, ErrUnlockTimeout.Default()
